Preventing two commands to be executed at the same time
There is an edge case when the filename is being edited and a file is 
being requested from the board. 

Both things will try to execute commands over serial on the board at the 
same time.

There should be a stack/queue of commands to be executed sequentially 
but this is not in the specification. 

Meanwhile the filename saving operation will take priority.