Add now required `teamId` option to notarization script
The options for the `notarize` call in the script are different depending on the setting of the `tool` option. Now that
the `tool` has been changed from the default `legacy` to `notarytool`, there is an additional `teamId` option, which is
the Apple Developer Program team ID associated with the macOS signing certificate.

Following the convention established in the Arduino IDE, I used an encrypted repository secret to configure the team ID
value. This is not absolutely necessary because, unlike some of the other credentials used by the build system, the team
ID is not actually a secret and can be seen in plaintext by anyone who examines the notarized application.